setup canonical URLs
We need defined canonical URLs as each actual HTML page is served from index.html in various directories.

The implementation for org pages is rather horrible as each org row in the country quries was previously unaware of the country urls defined by countries.rq. To make each row aware of this each country generator now duplicates the country URL/"safe name" information.

This issue should maybe be solved upstream in Snowman by 1. exposing the path in which the page renders in 2. global queries which are accessibe from all templates(similar to metadata)