✨ Add page for tools
AS a start only the browser extension is listed, but this can serve as a
place for more tools.
I only added the Firefox link, we can figure out later how to add the
Chrome one.
No link to this page yet as I couldn't find a good place to link from.

Fixes #209.